@@ -1,6 +1,6 @@
 {
   "name": "zotero-better-bibtex",
-  "version": "5.1.153",
+  "version": "5.1.160",
   "description": "Make Zotero useful for us LaTeX holdouts.",
   "homepage": "https://retorque.re/zotero-better-bibtex",
   "license": "ISC",
@@ -51,12 +51,12 @@
     "test": "test"
   },
   "dependencies": {
-    "@retorquere/bibtex-parser": "^1.0.75",
+    "@retorquere/bibtex-parser": "^2.0.0",
     "@types/lokijs": "^1.5.2",
     "ajv": "^6.10.2",
     "ajv-keywords": "^3.4.1",
     "barcoder": "^2.0.1",
-    "edtf": "^2.7.0",
+    "edtf": "^2.7.1",
     "edtfy": "^0.0.8",
     "escape-string-regexp": "^2.0.0",
     "eventemitter4": "^0.2.4",
@@ -76,9 +76,9 @@
     "kuroshiro": "^1.1.2",
     "kuroshiro-analyzer-kuromoji": "^1.1.0",
     "lokijs": "^1.5.8",
-    "njstrace": "^1.1.2",
+    "njstrace": "^1.2.0",
     "node-eta": "^0.9.0",
-    "parse5": "^5.1.0",
+    "parse5": "^5.1.1",
     "pegjs": "^0.10.0",
     "permutater": "^1.0.2",
     "punycode": "^2.1.1",
@@ -86,15 +86,15 @@
     "shelljs": "^0.8.3",
     "string-template": "^1.0.0",
     "transliteration": "^2.1.7",
-    "tslint": "^5.20.0",
-    "unicode2latex": "^1.0.23",
+    "tslint": "^5.20.1",
+    "unicode2latex": "^1.0.25",
     "webpack-cli": "^3.3.10",
     "xregexp": "^4.2.4",
-    "zotero-plugin": "^0.0.91"
+    "zotero-plugin": "^0.0.95"
   },
   "devDependencies": {
-    "@types/node": "^12.12.3",
-    "@types/webpack": "^4.39.7",
+    "@types/node": "^12.12.7",
+    "@types/webpack": "^4.39.8",
     "archiver": "^3.1.1",
     "bluebird": "^3.7.1",
     "circular-dependency-plugin": "^5.2.0",
@@ -105,7 +105,7 @@
     "ejs": "^2.7.1",
     "find-up": "^4.1.0",
     "fs-extra": "^8.1.0",
-    "glob": "^7.1.5",
+    "glob": "^7.1.6",
     "ignore-loader": "^0.1.2",
     "indent-string": "^4.0.0",
     "jsesc": "^2.5.2",
@@ -117,13 +117,13 @@
     "properties-reader": "^0.3.1",
     "pug": "^2.0.4",
     "request": "^2.88.0",
-    "request-promise": "^4.2.4",
+    "request-promise": "^4.2.5",
     "rimraf": "^3.0.0",
     "sorted-json": "^0.2.6",
     "ts-loader": "^6.2.1",
-    "ts-node": "^8.4.1",
+    "ts-node": "^8.5.0",
     "typedoc": "^0.15.0",
-    "typescript": "^3.6.4",
+    "typescript": "^3.7.2",
     "uri-templates": "^0.2.0",
     "webpack": "^4.41.2",
     "wrapper-webpack-plugin": "^2.1.0",
