@@ -12,7 +12,7 @@ declare const AddonManager: any
 import * as log from './debug'
 import { flash } from './flash'
 import { Events } from './events'
-import { ZoteroConfig } from './zotero-config'
+// import { ZoteroConfig } from './zotero-config'
 
 log.debug('Loading Better BibTeX')
 
@@ -30,7 +30,7 @@ import format = require('string-template')
 
 import { patch as $patch$ } from './monkey-patch'
 
-const prefOverrides = require('../gen/preferences/auto-export-overrides.json')
+import * as prefOverrides from '../gen/preferences/auto-export-overrides.json'
 
 /*
   UNINSTALL
@@ -256,7 +256,7 @@ Zotero.Translate.Export.prototype.Sandbox.BetterBibTeX = {
   parseHTML(sandbox, text, options) { return HTMLParser.parse(text.toString(), options) },
   extractFields(sandbox, item) { return varExtract(item) },
   debugEnabled(sandbox) { return Zotero.Debug.enabled },
-  version(sandbox) { return { Zotero: ZoteroConfig.Zotero, BetterBibTeX: require('../gen/version.js') } },
+  // version(sandbox) { return { Zotero: ZoteroConfig.Zotero, BetterBibTeX: require('../gen/version.js') } },
 
   debug(sandbox, prefix, ...msg) { Logger.log(prefix, ...msg) },
 
@@ -311,7 +311,7 @@ Zotero.Translate.Export.prototype.Sandbox.BetterBibTeX = {
 }
 Zotero.Translate.Import.prototype.Sandbox.BetterBibTeX = {
   debugEnabled(sandbox) { return Zotero.Debug.enabled },
-  version(sandbox) { return { Zotero: ZoteroConfig.Zotero, BetterBibTeX: require('../gen/version.js') } },
+  // version(sandbox) { return { Zotero: ZoteroConfig.Zotero, BetterBibTeX: require('../gen/version.js') } },
   parseHTML(sandbox, text, options) { return HTMLParser.parse(text.toString(), options) },
   platform(sandbox) { return Zotero.platform },
   debug(sandbox, prefix, ...msg) { Logger.log(prefix, ...msg) },
