@@ -3,7 +3,7 @@ declare const Zotero: any
 import * as log from './debug'
 import { Events } from './events'
 
-const defaults = require('../gen/preferences/defaults.json')
+import * as defaults from '../gen/preferences/defaults.json'
 const supported = ['debugLogDir', 'removeStock', 'postscriptProductionMode'].concat(Object.keys(defaults))
 
 // export singleton: https://k94n.com/es6-modules-single-instance-pattern
