@@ -5,9 +5,10 @@ import { upgradeExtra } from './upgrade-extra'
 // import { getItemsAsync } from '../get-items-async'
 import { DB as Cache } from './cache'
 import * as log from '../debug'
+import { flash } from '../flash'
 
 export async function upgrade(progress) {
-  progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n: '?', total: '?' })) // tslint:disable-line:no-magic-numbers
+  progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n: '?', total: '?' }))
 
   const patterns = []
   for (const prefix of ['bibtex', 'biblatexcitekey', 'biblatex', 'biblatexdata']) {
@@ -28,6 +29,7 @@ export async function upgrade(progress) {
       AND (${patterns.map(pattern => 'itemDataValues.value like ?').join(' OR ')})
   `
 
+  let notEditable = 0
   await Zotero.DB.executeTransaction(async () => {
     const legacy = await ZoteroDB.queryAsync(query, patterns)
     if (!legacy.length) return
@@ -35,7 +37,7 @@ export async function upgrade(progress) {
 
     const total = legacy.length
     let n = 0
-    progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n, total })) // tslint:disable-line:no-magic-numbers
+    progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n, total }))
 
     for (const item of legacy) {
       n += 1
@@ -45,19 +47,34 @@ export async function upgrade(progress) {
       if (extra !== item.extra) {
         log.debug('dbUpgrade: old=\n', item.extra)
         log.debug('dbUpgrade: new=\n', extra)
+
         const upgraded = await Zotero.Items.getAsync(item.itemID)
-        try {
-          upgraded.setField('extra', extra)
-        } catch (err) {
-          await upgraded.loadAllData()
-          upgraded.setField('extra', extra)
+
+        if (!upgraded.isEditable()) {
+          log.error('dbUpgrade:', item.extra, 'required upgrade, but', item.itemID, 'is not editable')
+          notEditable++
+
+        } else {
+          try {
+            upgraded.setField('extra', extra)
+          } catch (err) {
+            await upgraded.loadAllData()
+            upgraded.setField('extra', extra)
+          }
+          await upgraded.save()
+          affected.push(upgraded.id)
         }
-        await upgraded.save()
-        affected.push(upgraded.id)
       }
     }
-    progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n: total, total })) // tslint:disable-line:no-magic-numbers
+    progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade', { n: total, total }))
     if (affected.length) Cache.remove(affected, 'dbUpgrade')
   })
-  progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade.saving')) // tslint:disable-line:no-magic-numbers
+  progress(Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade.saving'))
+
+  if (notEditable) {
+    flash(
+      Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade.saving'),
+      Zotero.BetterBibTeX.getString('BetterBibTeX.startup.dbUpgrade.notEditable', { n: notEditable })
+    )
+  }
 }
