@@ -52,7 +52,7 @@ export let KeyManager = new class { // tslint:disable-line:variable-name
         if (inspireHEP) {
           parsed = varExtract({ extra: item.getField('extra') })
 
-          let key = parsed.extraFields.csl.DOI || item.getField('DOI') || arXiv.parse(parsed.extraFields.kv.arxiv).id
+          let key = parsed.extraFields.csl.DOI || item.getField('DOI') || arXiv.parse(parsed.extraFields.tex.arxiv).id
           if (!key && ['arxiv.org', 'arxiv'].includes((item.getField('libraryCatalog') || '').toLowerCase())) key = arXiv.parse(item.getField('publicationTitle')).id
           if (!key) throw new Error(`No DOI or arXiv ID for ${item.getField('title')}`)
 
