@@ -54,6 +54,8 @@ function __njsOnCatchClause__(call) {
 }
 `
 
+const logArguments = false
+const logExceptions = true
 export = function loader(source) {
   const filename = this.resourcePath.substring(process.cwd().length + 1)
 
@@ -63,5 +65,5 @@ export = function loader(source) {
 
   console.log(`!!!!!!!!!!!!!! Instrumenting ${filename} for trace logging !!!!!!!!!!!!!`)
 
-  return tracer + injector.injectTracing(filename, source, true, true)
+  return tracer + injector.injectTracing(filename, source, logExceptions, logArguments)
 }
