@@ -186,10 +186,8 @@ template = """
 
 declare const Zotero: any
 
-// don't take this from Translator.isZotero because that initializes after library load
-const client = Zotero.BetterBibTeX.version().Zotero
-const zotero = client.isZotero
-const jurism = client.isJurisM
+const jurism = Zotero.Utilities.getVersion().includes('m')
+const zotero = !jurism
 
 export const valid = new Map([
   #for $typeName, $fields in $validFields.items()
