@@ -45,7 +45,7 @@ const common = {
   resolveLoader: {
     alias: {
       'pegjs-loader': 'zotero-plugin/loader/pegjs',
-      'json-jsesc-loader': 'zotero-plugin/loader/json',
+      // 'json-jsesc-loader': 'zotero-plugin/loader/json',
       'bcf-loader': path.join(__dirname, './setup/loaders/bcf.ts'),
       'trace-loader': path.join(__dirname, './setup/loaders/trace.ts'),
     },
@@ -53,7 +53,7 @@ const common = {
   module: {
     rules: [
       { test: /\.pegjs$/, use: [ 'pegjs-loader' ] },
-      { test: /\.json$/, type: 'javascript/auto', use: [ 'json-jsesc-loader' ] }, // https://github.com/webpack/webpack/issues/6572
+      // { test: /\.json$/, type: 'javascript/auto', use: [ 'json-jsesc-loader' ] }, // https://github.com/webpack/webpack/issues/6572
       { test: /\.bcf$/, use: [ 'bcf-loader' ] },
       { test: /\.ts$/, exclude: [ /node_modules/ ], use: [ 'trace-loader', 'ts-loader' ] },
     ],
